Enhance image uploading functionality: update division and tour services to handle image uploads and deletions, integrate multer for file handling, and update user service to manage user images.

+++ src/app/modules/division/division.service.ts
@@ -37,7 +37,11 @@ const updateDivisionByID = async (id: string, data: Partial<IDivision>) => {
     runValidators: true
   })
 
-  if (isDivisionExist.thumbnail) {
+  if (
+    isDivisionExist.thumbnail &&
+    data.thumbnail &&
+    isDivisionExist.thumbnail !== data.thumbnail
+  ) {
     await deleteImageFromCloudinary(isDivisionExist.thumbnail)
   }
